requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
shap==0.48.0
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
ccxt==4.5.12
pyarrow==17.0.0
